Major code refactor
- Sorry Mars I pushed too many changes into a single commit
- Moved to non-flat directory structure to enable future tests and to make installing easier
- Fixed issues with 2nd order models not having the correct length of time vectors
- Moved notebooks to their own directory
- Moved results to be under the notebook directory
- Tested the 2nd order representations (notebooks/synthetic_data_analysis_detailed.2nd-order). SINDy-SHRED
really struggles when the latent space velocity is strongly non-linear (apparently?)
- Cleaned up the MIOSR affiliated tools a bit
- Started experimenting with an autotune function for MIOSR model discovery.
