+++ src/sindy_shred/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    "sindy",
+    "sindy_shred",
+    "miosr_utils",
+    "processdata",
+    "models",
+    "plotting",
+]
+
+
+from .sindy_shred import SINDySHRED
